pyproject.toml
[tool.poetry]
name = "suinfra-cli"
version = "0.1.0"
description = ""
authors = ["Studio Mirai <contact@sm.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
bunnet = "^1.3.0"
pysui = "0.60.0"
python-dotenv = "^1.0.1"
trio = "^0.25.1"
typer = "^0.12.3"

[tool.poetry.scripts]
suinfra = "suinfra_cli.main:app"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
